"Fix" unit tests.
Note that the later versions of pinexq-client use a different syntax for create_and_configure_rapidly().

The statevectors produced by the circuits are *physically* correct, but they have the wrong phase.
(And why is the overlap almost exactly 1j, *sometimes*?)
That isn't an SDK responsibility, though. We need to check the implementation...