Retrieve and fix the (newly available) global phase for both Qiskit and Pennylane integration.
For Qiskit, we simply need to set the "global phase" of the QuantumCircuit.
For Pennylane, we add a GlobalPhase gate to the start of the operations list.

Notably, since find_processing_step in initialize.py can now find the most recently created step, we do not need to update that part.