Add support for build_initialization_circuits, which can prepare multiple states at once.
This feature is not accessed through QAlchemyInitialize, since that class extends Gate.
This feature also necessitated further changes in job submission and retrieval.
(Including adding a parameter that was accidentally not set optional on baa_procon.)

See test_initialize.py for a simple test of this new feature.

Note: Currently, the batch function can only accept a _list_ of statevectors.
We may wish to also support 2D arrays/matrices...