Add and test QASM3 and other extra_kwargs. Also fix an issue with the global phase.
More precisely, the issue is that baa_procon does not record global phase if QASM3 is on;
in that case, loading the QASM3 sets the quantum circuit's global phase, so trying to "fix" it like in the QASM2 case zeroes it out.

I can probably fix that on the baa side later, but it's not really necessary anyway.